active_trader_llm/requirements.txt
# Core dependencies for ActiveTrader-LLM

# Python version: 3.11+

# Data handling
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0

# Market data
yfinance>=0.2.28
alpaca-py>=0.20.0

# Technical indicators
tradingview-indicators>=1.0.0

# LLM APIs
anthropic>=0.18.1
openai>=1.10.0
google-generativeai>=0.3.0  # For Gemini models (optional)
requests>=2.31.0  # For HTTP-based providers (xAI, DeepSeek, OpenRouter)

# Configuration
pyyaml>=6.0

# Database
sqlalchemy>=2.0.25

# Utilities
python-dotenv>=1.0.0
pytz>=2024.1

# Optional: Web API (for paper-live mode)
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
